Simplifies the AddEditTask RIB.
The presenter has only one method to pass to the view the task (or empty
task) to be edited. If the task passed as parameter is Task.EMPTY that
would mean we need to create a new task.

The method editOrAddTask() returns an Observable<Task> which emits when
the view completes editing or adding the new task.

Test are also updated.